refactor: rename Data Ingestion tab and component to Overview
