--- frontend/src/views/dataIngestion/index.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import Container from 'react-bootstrap/Container'
-
-import IngestionStatistics from './stats'
-
-const DataIngestion = () => {
-  return (
-    <Container>
-      <IngestionStatistics />
-    </Container>
-  )
-}
-
-export default DataIngestion
